ZCS-8309: Fix for File Upload not working. (#278)
There was 2 problems related to this bug"
1. File Upload was not working.
2. It was not Honoring Affinity.

On further analysis it was found out that it was not able to upload the attachments when the request is going to the other pod to upload the attachment rather than on the pod where the temporary upload_XXXX:XXXX.tmp and the other metadata file uploadmetadata_XXXX:XXXX.tmp is getting generated. So, it was not able to fetch the upload and was reporting the NO_SUCH_UPLOAD exception. It was found that this behaviour was
very frequently getting observed when we are enabling the 'Send As' and 'Send on behalf of' feature where when we are granting the other user to 'Send As' or 'Send on behalf of' when they are existing on the other pod. The reason was that the uploadId is constructed from serverId and the uuid of the attachment file and from this uploadId we were fetching the serverId where the file is getting generated and in case of ZimbraX it will always be the serverId of the local server so it was failing when the file is getting generated on the other pod. To fix this I have reconstructed the uploadId and added the IP at the end of it. Now, the uploadId consist of 3 parameters rather than the older 2 which are now serverId:uuid:IP. So, now I am fetching this IP from uploadId and providing it to the URI after checking if it's a local upload or a remote upload.

The other problem was observed when we were trying to send the e-mail if the attachment were successful, so it was reporting WRONG_HOST exception which was coming because of the method isLocal() which was deciding whether it is on the same host or the remote host. So, again it was deciding on the basis of the hostname and in case of ZimbraX it will always be 'zmc-mailbox'. Now, this is replaced with the IP based comparison to detect whether it's a local pod or a remote pod for which the request needs to be proxied.