ProxyConfGen zimbra_admin upstream routes to zmc-mls
The old WebAdminUpstreamServersVar class used for populating
the zimbra_admin upstream directive used the names of all
mail servers, which resolved to "zmc-mailbox". This caused admin
SOAP requests to bypass the mls, breaking affinity.
By replacing it with the existing WebAdminUpstreamAdminClientServersVar
used for populating zimbra_adminclient, the issue is resolved.