correct handling of server proxy target attribute
If a SOAP request has a "targetServer" value, then we actually
need to use a ServerProxyTarget instead of an IpProxyTarget.
Currently, this will always result in the request NOT being
proxied, since only one "server" exists (zmc-mailbox), but
it's still good to have since clients can still use this header.
It also future-proofs us if we ever move to an architecture that
has multiple logical "servers" represented by different pod groups.