ZCS-8223: Refactored the proxyRequest method to (#290)
Make use of the affinity IP derived from the account ID. Before that, while creating the galSync account and if it should be proxied then it was not able to get the correct IP because of that change IllegalArgumentException was getting observed. To fix that correct IP address of the pod is derived from the zimbraID and changed the proxyRequest method to make use of that.