ZCS-8748:Fix for permission denied:not an eligible
admin account.

This excpetion was coming only in the case when the
two accounts where on the two different pods and it
was working fine in case if the two accounts are on
the same mailbox pod. Also, there is no problem got
observed accessing the public sharing calendar with
the use of the link provided the user who wants to
share its calendar publicly.

Using the remote debugging onto my AWS instance and
found that the ownerId was getting supplied for the
wrong account from GetShareInfo.fetchRemoteShareInfo
method. And it was getting diverted to the wrong
DocumentHandler.proxyRequest method. Now, changed it
to the earlier way of getting server replaced with the
way of getting IP.