ZCS-8854: Fix for broken external sharing. (#306)
External sharing was not working in the ZimbraX as on sharing the folder externally it is found out that the link which was getting shared externally for registering as well as for accessing the account were using the podIP of the mailbox compared to earlier way of using hostname
of the server. To fix this, if the publicservicehostname is configured in the admin console then it will use that and build the link for registering and accessing account and if it is not set then it will use the older way of building the link.

And Addressed review comments.