ZCS-8846: Fix for the broken RestURL. (#307)
RestURL was not correct for user with domain having zimbraPublicServiceHostname/zimbraPublicServicePort. So, on further analysis found that because of the recent changes made to affinity in the REST URL mailbox Pod IP was coming instead of the required zimbraPublicServiceHostname. So, made changes to fetch the zimbraPublicServiceHostname if configured and constitute the REST URL.

And addressed review comments.