ZCS-8412:Fixed Shared calendar item entry not show
Shared calendar item entry was getting shown on UI after a
CreateAppointmentRequest/ModifyAppointmentRequest and the
CancelAppointmentRequest. It was only shown when the
BatchRequest comprising of the SearchRequest and the
GetMiniCalRequest fires up. This BatchRequest is generally
called after every time after CreateAppointmentRequest or
ModifyAppointmentRequest or CancelAppointmentRequest fires.
This issue only pops up in the old web-client not in the
new web-client. As new web-client uses other mechanism to
fetch the updated entries and shows up in the UI. To fix
this issue it is found that handleRemoteNotifications method
in the SoapSession class is having issues on the referesh.
If there is a remote refersh it should call the overall refersh
but it was not happening, on futher analysing it is found out
that when Element is tring to find ZimbraNamespace.E_REFRESH is
not set to null the condition fails over there because it is
never set. To address that in The DocumentHandler.proxyWithNotification
method while handling the soap session and checking is the
notifications are enabled then ZimbraNamespace.E_REFRESH is
set to the ELement, which enables the refersh in the old web-client
on CreateAppointmentRequest/ModifyAppointmentRequest/CancelAppointmentRequest.