ZCS-8840: Fix for failing DeleteGalSyncAccountRequest (#324)
Fixed the failing DeleteGalSyncAccountRequest while sanity check. After creating the domain and issuing the CreateGalSyncAccountRequest successfully now when it tries to issue a DeleteGalSyncAccountRequest then sometimes when it lands on the correct pod it get executed successfully but when it lands on the wrong pod it complains for NO_SUCH_ACCOUNT. On analysing the codebase it is found that there is no proxying mechanism written for the DeleteGalSyncAccountRequest so whenever it lands on the wrong pod then this issue pops up. So, to fix that written the logic for proxying to the correct pod.