ZCS-8845: Fix for permissions_guest.xml test case (#327)
zm-soap-harness/data/soapvalidator/RestServlet/Sharing/permissions_guest.xml this REST API test case was failing because when it is trying to perform the basic authentication and when the server is challenged then it should add the Authorization header in the request. As, there was no REST logs can be enabled to trace this used the remote debugging on the suspected UserServlet class and found that it was failing at the AuthUtil class when it is trying to retrieve
the HTTP_AUTH_HEADER header in the request and when it is not found then the above mentioned test case is failing and saying for must-authenticate. To fix this problem a condition has been added if the Authorization header is not found then it will fallback to WWW_AUTHENTICATE_HEADER on the server side and perform the authentication where the realm is provided as zimbraBasicAuthRealm.