determine pod function based on name
Instead of using the is_mailbox_pod LC key, instead look at
the pod name (as provided by the k8s downward API) to determine
if the current pod is a zmc-backup-restore pod.
This is because we don't currently have a simple way of configuring LC values
for *just* zmc-backup-restore pods.