ZBUG-1456 Wrap early request in batch for getInfo if accountId/infoResult are not set.
If accountId and mGetInfoResult are not set, convert to a batch request to fetch
them in addition to whatever request we are already running so the first round of
refresh handling does not fail due to missing accountId. This is necessary to
properly handle folder caching which requires an accountId.