ZBUG-1456 (#1104)
* ZBUG-1456 Connection pool shut down 3.0

* ZBUG-1456 Wrap early request in batch for getInfo if accountId/infoResult are not set.

If accountId and mGetInfoResult are not set, convert to a batch request to fetch
them in addition to whatever request we are already running so the first round of
refresh handling does not fail due to missing accountId. This is necessary to
properly handle folder caching which requires an accountId.

* ZBUG-1456 Add a bodyHandler callback to invoke method to allow callers to
process the body before the context is processed.

Update the auth methods to use the bodyHandler to set appropriate auth
information before handling context. This is useful to prevent failures
due to invalid auth if notifications are received in the init auth request.

* ZBUG-1456 Disable alwaysRefreshFolders for modify events.