Fix for alias expansion for members of "foreign" domain distribution lists during domain rename
Among the phases of a domain rename there is one where distribution
lists of domains not involved in the rename procedure, but some of whose
members belong to the domain which does, see the domain of those members
changed to the target domain of the rename operation.

Unfortunately if one of those members is an address of an account with
one or more aliases, what happens is that the address of the old domain
is removed from the distribution list but instead of being replaced by
the address with the same localpart in the new domain, it is replaced by
all the addresses of the account belonging to the new domain.

This commit corrects that behaviour.