TSS-18004: Fix for PTD - Message not displaying correctly
Analysis:
It was reported that some of the messages were not getting
displayed correctly. On further analysis it was found that
the particular attached message was having a lot of malformed
inline styling in the original message body which contains a
lot of unclosed double quotes which was escaping large chunks
of the real message. The mailer which customer has used seems
like sendgrid which has malformed the html.

Looking through the Antisamy policy which we are using turned
out that it cannot be fixed through that the reason is that the
antisamy policy which we are using with owasp can able to handle
various operations like validate, truncate, filter, remove based
on the element types, tag names, tag values, attributes and the
css rules but not with the malformed qoutes.

Looking inside the owasp java sanitizer code turns out that the
owasp uses HtmlLexer for the sanitization process which traverses
element by element so before sanitizing the html it takes the
value of the attribute and strips of any vulnerable stuffs based
on what policies we have applied. But in our case the problem is
not that OWASP html filter is strict with the quotes rather it is
skipping the texts because it is considering the text as the
attribute name because of the malformed double quotes which was
present inside the span tag and it thinks of it as the new
attribute has started and interepts whatever is there as its value
before the next double quotes.

For example in case of the following html:
<span style="font-size:11pt;font-family:Arial;font-variant-ligatures:normal;font-variant-east-asian:normal;font-variant-position:normal;vertical-align:baseline"">Dear Member,</span></p><br style="-webkit-text-size-adjust:auto;" />

The attribute style is properly interpreted then the next malformed
double  quote it has interpreted as the !attrsReadyForName and it
has interpreted the key as: ">Dear Member,</span></p><br style="
and detected its value as -webkit-text-size-adjust:auto;" and since
the value parts looks vulnerable to owasp so it has quitely dropped.

And as per comments from the owasp community "Unfortunately we cannot
throw compilation errors for invalid HTML markup."

On anlysing the already reported issues of antisamy and owasp it is
found that user has to handle the invalid HTML markup.

Fix:
Looked some libraries that can fix the malformed html for that tried
using jsoup which is used for parsing, cleaning, extract, maipulate
and make the output tidy. Used its cleaner class and got the expected
results. Later it is realised that jsoup is also capable of doing the
sanitazation so to remove any future collision between the two jsoup
is dropped.

Finally, used the java port of the old "HTML Tidy" which is originally
written in C and is a command line utility which is widely used for the
cleaning the malformed markup. So, used the java port JTidy which is
updated to work with HTML5 new tags. Used its Tidy class with other
options for cleaning the malformed html. The advantage of using this
library is that it only fixes the markup but does not do any other
process of sanitization or manipulation as jsoup does. The cleaned
html is now passed to the sanitize() method for further sanitization
process using antisamy and owasp.