ZCS-10322: Added custom-mimetypes.xml in
store-conf which will be copied over to the
/opt/zimbra/conf directory where we can be
able override the existing mimetypes and
can be able to map the content-type to any
of the extensions with a higher magic priority.
So, in this change we are using compositeDetector
which comprises of the default detector as well
as the custom detector. If the custom-mimetype.xml
is missing then in that case the mediaType will
fallback to the default detector and will use the
default tika-mimetypes.xml file.