ZCS-10322: Removed the older way of contentType
detection using the MimeDetect class and removed
the unnecessary code for handling the .har file
only based on the extension. As, if we remove the
extension of the .har file ZBUG-711 still exist
and making the web-client unresponsive and also
in CPU spike. While it is correctly handled by
the tika with it's content-type detected as
application/pdf contraray to earlier where it is
getting detected as application/octet-stream but
the file actually contains html content and the
convertd is trying to convert the document based
on 'application/octet-stream' and failing.