ZCS-10443: Modified zimbraFileUploadMaxSizePerFile
ldap attribute to be used for file uploads restrictions
other than e-mail. Now, the old defaultCOSValue is
removed as it's value should be less than the ldap
attribute zimbraFileUploadMaxSize. If the default
value is not set then upload size is limited by
zimbraFileUploadMaxSize which is global.
Also, added accountInfo flag to zimbraFileUploadMaxSizePerFile.