TSS-18300: Fix for emails displaying as blank
It was observed that certain e-mails were displaying
as blank. On analysing the input mime it was found
that there was issue with the input html that there
was no closing "style" tag. Here, there are 2 issues:

1. In the JTidy configuration setPrintBodyOnly() was
earlier set to true so it was processing only the
body part and skipping the style tags.

2. Earlier in the code skipped the trimming of the
parsed string which was getting null value as it
was not handling the head of the html because of
setPrintBodyOnly() was set to true but the code  was
written to handle these situation earlier and fallback
to the owasp if the output from the JTidy was null. In
this particular case it was not able to handle because
the parsed string was having a whitespace character and
it was not null so it skipped the string trimming which
is causing not getting fallback to owasp directly.