ZCS-10443: Added new ldap attributes for restricting
the maximum size in bytes for e-mail attachment zimbraMailAttachmentMaxSize
and the for file uploads restrictions zimbraFileUploadMaxSizePerFile other
than e-mail. The default value for both the new ldap attributes is set to 0.
So, the provided file upload size limit to account and COS levels. As we are
using the zimbraMailAttachmentMaxSize ldap attribute for e-mail attachments
file size restriction and zimbraFileUploadMaxSizePerFile for the briefcase
and other file size restrictions. And if these new ldap attributes are not
set then it falls back to the original ldap values zimbraMtaMaxMessageSize
in case of e-mail attachment file size restriction else zimbraFileUploadMaxSize.
The values of these new attributes are lower than the older global level
ldap attributes (zimbraMtaMaxMessageSize, zimbraFileUploadMaxSize). If the
value of (zimbraMailAttachmentMaxSize, zimbraFileUploadMaxSizePerFile) is set
higher than the (zimbraMtaMaxMessageSize, zimbraFileUploadMaxSize) then the
global value is overriden.