TSS-18349: Emails from Expedia not displaying
It was observed that certain e-mails were displaying as blank.
On analysing the input mime it was found that there was issue
with the input html after getting parsed div tag was containing
a new line between "<" and "/div>". To fix this issue a new
condition is troduced to handle if any of the end tag contains
this situation then newline is removed from that tag and it
should constitute a well-formed end tag.