ZCS-10544/TSS-18349: Fix for emails displaying as blank.
It was observed that certain e-mails were displaying as blank.
On analysing the input mime it was found that there was issue
with the input html after getting parsed div tag was containing
a new line between < and /div>. Handled the newline condition
using the JTidy library option setWraplen(0) as earlier JTidy
was wrapping the html at 68th character. So, setting this option
to set the wrapping to 0 which in removes any chances of having
the unnecessary newline in the html.

Also, added localconfig hide_jtidy_warnings to disable the printing
of warnings and messages in the zmmailboxd.out log. Also, added the
following configs:

Set to omit the doctype to not have unnecessary jtidy generated
doctype in the parsed html. Set the tidy mark to false so that
tidy should not add meta element indicating tidied doc. Set the
hide end tags to true so that it suppress optional end tags.