ZBUG-1948: Fix for 427353-1804751.msg because of
java.lang.IndexOutOfBoundsException

The root cause of this problem is that the attribute value in
itself is "background" instead of attribute itself, so while
BackgroundPolicy tries to seek the attrs.indexOf("background")
resulting in wrong index of the attribute and resulting in the
java.lang.IndexOutOfBoundsException.

To fix this another condition is implemented which checks for
that if the keyword "background" is coming at odd index or not,
if the index of "background" keyword is found at the odd index
postion then we are ignoring it as we are only interested in
the attribute key if it is "background" or not.