ZCS-9764: Fix for SSRF Vulnerability in preauth servlet.
As per description in the ticket an authenticated user can
recover his authoken by simply looking at what URLs are
loaded by his browser after authentication. This valid
authtoken can be used to exploit the insecure redirect in
the preauth servlet

/service/preauth? isredirect=1&authtoken=[VALIDTOKEN]&redirectURL=http://whatever

To fix the above exploitation checked the redirectURL in the
PreAuthServlet to look if the domain is localhost or 127.0.0.1
or it belongs to the set of private IP addresses then it restricts
the user from redirecting it to the local addresses and if it is an
external addresses then it allows the redirection.