ZCS-9764: Fix for Open Redirect Vulnerability in PreAuth servlet
There is an insecure redirect in the preauth servlet which
can only be exploited with a valid authtoken as compared to
earlier fix this doesn't hold true as a SSRF Vulnerability.
This will allow redirect to any external url without checking
a whitelist. User can also specify an internal port, which is
not specified publicly exposed, which will allow user to probe
internal ports.

To fix this, in the redirectURL if it contains the protocol with
a domain name and/or a port then that part of the url is stripped
off from the redirectURL and restricting it from getting access to
any of the internal domain and ports.