TSS-18004: Updated owasp-java-html-sanitizer to 20190610.2z
It was found that the particular attached message was having
a lot of malformed inline styling in the original message body
which contains a lot of unclosed double quotes which was escaping
large chunks of the real message. To fix this issue the code
changes has been done in the owasp-java-html-sanitizer and it's
jar file is updated in the ivy.xml.