ZBUG-2167:SMIME signature corrupted in Japanese mail (#1163)
[Problem]
When the message's subject and/or body is written in Japanese and its message structure is plain/text, the S/MIME signed message is corrupted and appeared as "tampered" in the recipient Inbox.

[Root cause]
When the message was a single part message (plain/text), the message was first sent to the MBS in raw UTF-8 characters, then MBS performed the S/MIME-sign operation to that raw UTF-8 email. But after that MBS re-checked the message body and encoded the raw UTF-8 to Quoted printable.  In short, MBS tampered the message.

[Fix]
Before S/MIME-sign operation, MBS encodes the raw UTF-8 message into the Quoted printabble one, then S/MIME-signs it.

[Manual testing]
(1) Setup the S/MIME configuration for two accounts
(2) Compose a plain/text message with a Japanese subject and Japanese body data, sign it with S/MIME and send it.
(3) Verify that the message is delivered with "signed by xxx" message in the header area.