ZCS-11022: Implemented CSRF validation
For 2FA enabled admin account, csrfToken is returned in response
of 1st AuthRequest with username/password. That csrfToken is now
sent in 2nd AuthRequest with TOTP. After 2nd AuthRequest succeeds,
the csrfToken returned is used in subsequent soap requests and is
validated against the autoken for its validity.

If CSRF is disabled then, csrfToken is not returned in in the
response of 1st AuthRequest with username/password.